mensaje de confirmacion agregado al eliminar producto

+++ frontend/src/components/Products.jsx
@@ -1,18 +1,48 @@
 import Product from "./Product"
 import { AdminContext } from "../../context/adminContext"
 import { useContext } from "react"
+import Swal from "sweetalert2";
 
 const Products = ({ products }) => {
 
-    const {admin, deleteProduct, setProducts} = useContext(AdminContext)
+    const { admin, deleteProduct, setProducts } = useContext(AdminContext)
 
     const handleDelete = async (id) => {
-        try {
-            await deleteProduct(id);
-            const newProducts = products.filter((product) => product._id !== id);
-            setProducts(newProducts);
-        } catch (error) {
-            console.log(error);
+       
+        const result = await Swal.fire({
+            title: "¿Estás seguro de eliminar este producto?",
+            text: "No podrás revertir esta acción",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#d33",
+            cancelButtonColor: "#3085d6",
+            confirmButtonText: "Sí, eliminar",
+            cancelButtonText: "Cancelar"
+        });
+
+        if (result.isConfirmed) {
+            try {
+                await deleteProduct(id)
+                const newProducts = products.filter((product) => product._id !== id)
+                setProducts(newProducts)
+
+                
+                Swal.fire({
+                    title: "Eliminado",
+                    text: "El producto fue eliminado correctamente",
+                    icon: "success",
+                    confirmButtonColor: "#3085d6"
+                })
+
+            } catch (error) {
+                console.error(error)
+                Swal.fire({
+                    title: "Error",
+                    text: "Ocurrió un problema al eliminar el producto",
+                    icon: "error",
+                    confirmButtonColor: "#3085d6"
+                })
+            }
         }
     }
 
@@ -22,7 +52,7 @@ const Products = ({ products }) => {
 
                 {products?.map((product) => (
                     <div key={product._id}>
-                        <Product 
+                        <Product
                             id={product._id}
                             admin={admin}
                             productImage={product.image}
